connect login page, and cart to user route

+++ db/schema/03_menu_items.sql
@@ -3,7 +3,7 @@ CREATE TABLE menu_items (
   id SERIAL PRIMARY KEY NOT NULL,
   category_id INTEGER REFERENCES categories(id) ON DELETE CASCADE,
   name VARCHAR(255) NOT NULL,
-  price INTEGER NOT NULL,
+  price DECIMAL(5, 2) NOT NULL,
   image TEXT,
   description VARCHAR(500),
   option INTEGER
